requirements.txt
# File path: requirements.txt

# Core Libraries
playwright

# Configuration
python-dotenv
configparser

# Utilities
pypdf
langchain
fuzzywuzzy
python-Levenshtein

# Testing
pytest
pytest-mock